Moved Pointer-based array stuff to seperate file to clean up stuff
