+++ TestReadFile/TestReadFile/TestReadFile.vcxproj
@@ -84,12 +84,14 @@
     <ClInclude Include="concol.h" />
     <ClInclude Include="List.h" />
     <ClInclude Include="Music.h" />
+    <ClInclude Include="PtrArr.h" />
     <ClInclude Include="Utility.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="concol.cpp" />
     <ClCompile Include="List.cpp" />
     <ClCompile Include="Main.cpp" />
+    <ClCompile Include="PtrArr.cpp" />
     <ClCompile Include="Music.cpp" />
     <ClCompile Include="Utility.cpp" />
   </ItemGroup>
